+++ CHANGELOG.md
@@ -5,11 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [8.1.0] - unreleased
+## [9.0.0] - unreleased
 
 ### Fixed
 
 - Progress download column now displays decimal units
+
+### Added
+
+- Added legacy_windows to ConsoleOptions
+
+### Changed
+
+- Dropped box.get_safe_box function in favor of Box.substitute
 
 ## [8.0.0] - 2020-10-03